Fixed a bug in 4.3.1 that caused AI to be unable to use some weapons
Code edits regarding AIpref in Enums.h.

Added GamePids.h (derived from ItemPid.h in mapper scripts)

+++ sfall/Modules/Combat.cpp
@@ -282,7 +282,7 @@ static long __fastcall CheckDisableBurst(fo::GameObject* critter, fo::GameObject
 			return 10; // Disable Burst (area_attack_mode - non-existent value)
 		}
 	}
-	return cap->area_attack_mode; // default engine code
+	return (long)cap->area_attack_mode; // default engine code
 }
 
 static void __declspec(naked) ai_pick_hit_mode_hack_noBurst() {
